+++ packages/nx-plugin/src/executors/integration-test-suite/schema.d.ts
@@ -1,0 +1,6 @@
+export interface IntegrationTestSuiteExecutorSchema {
+  cwd: string;
+  testFilePath: string;
+  withRegistry?: boolean;
+  updateSnapshots?: boolean;
+}