+++ packages/schematics/src/utils.ts
@@ -231,10 +231,7 @@ export function setESLintProjectBasedOnProjectType(
 ) {
   let project;
   if (projectType === 'application') {
-    project = [
-      `${projectRoot}/tsconfig.app.json`,
-      `${projectRoot}/tsconfig.spec.json`,
-    ];
+    project = [`${projectRoot}/tsconfig.(app|spec).json`];
 
     if (hasE2e) {
       project.push(`${projectRoot}/e2e/tsconfig.json`);
@@ -242,19 +239,13 @@ export function setESLintProjectBasedOnProjectType(
   }
   // Libraries don't have an e2e directory
   if (projectType === 'library') {
-    project = [
-      `${projectRoot}/tsconfig.lib.json`,
-      `${projectRoot}/tsconfig.spec.json`,
-    ];
+    project = [`${projectRoot}/tsconfig.(lib|spec).json`];
   }
   return project;
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export function createRootESLintConfig(
-  prefix: string | null,
-  hasE2e?: boolean,
-) {
+export function createRootESLintConfig(prefix: string | null) {
   let codeRules;
   if (prefix) {
     codeRules = {
@@ -277,13 +268,9 @@ export function createRootESLintConfig(
     overrides: [
       {
         files: ['*.ts'],
-        parserOptions: {
-          project: hasE2e
-            ? ['tsconfig.json', 'e2e/tsconfig.json']
-            : ['tsconfig.json'],
-          createDefaultProgram: true,
-        },
         extends: [
+          'eslint:recommended',
+          'plugin:@typescript-eslint/recommended',
           'plugin:@angular-eslint/recommended',
           'plugin:@angular-eslint/template/process-inline-templates',
         ],
@@ -300,26 +287,29 @@ export function createRootESLintConfig(
 }
 
 function createProjectESLintConfig(
-  rootPath: string,
   projectRoot: string,
   projectType: ProjectType,
   prefix: string,
+  setParserOptionsProject: boolean,
   hasE2e: boolean,
 ) {
   return {
-    extends: `${offsetFromRoot(rootPath)}.eslintrc.json`,
+    extends: `${offsetFromRoot(projectRoot)}.eslintrc.json`,
     ignorePatterns: ['!**/*'],
     overrides: [
       {
         files: ['*.ts'],
-        parserOptions: {
-          project: setESLintProjectBasedOnProjectType(
-            projectRoot,
-            projectType,
-            hasE2e,
-          ),
-          createDefaultProgram: true,
-        },
+        ...(setParserOptionsProject
+          ? {
+              parserOptions: {
+                project: setESLintProjectBasedOnProjectType(
+                  projectRoot,
+                  projectType,
+                  hasE2e,
+                ),
+              },
+            }
+          : null),
         rules: {
           '@angular-eslint/directive-selector': [
             'error',
@@ -340,7 +330,10 @@ function createProjectESLintConfig(
   };
 }
 
-export function createESLintConfigForProject(projectName: string): Rule {
+export function createESLintConfigForProject(
+  projectName: string,
+  setParserOptionsProject: boolean,
+): Rule {
   return (tree: Tree) => {
     const angularJSON = readJsonInTree(tree, 'angular.json');
     const {
@@ -366,10 +359,10 @@ export function createESLintConfigForProject(projectName: string): Rule {
         : createRootESLintConfigFile(projectName),
       updateJsonInTree(join(normalize(projectRoot), '.eslintrc.json'), () =>
         createProjectESLintConfig(
-          tree.root.path,
           projectRoot,
           projectType,
           prefix,
+          setParserOptionsProject,
           hasE2e,
         ),
       ),
@@ -392,7 +385,6 @@ function createRootESLintConfigFile(projectName: string): Rule {
   return (tree) => {
     const angularJSON = readJsonInTree(tree, getWorkspacePath(tree));
     let lintPrefix: string | null = null;
-    const hasE2e = determineTargetProjectHasE2E(angularJSON, projectName);
 
     if (angularJSON.projects?.[projectName]) {
       const { prefix } = angularJSON.projects[projectName];
@@ -400,7 +392,7 @@ function createRootESLintConfigFile(projectName: string): Rule {
     }
 
     return updateJsonInTree('.eslintrc.json', () =>
-      createRootESLintConfig(lintPrefix, hasE2e),
+      createRootESLintConfig(lintPrefix),
     );
   };
 }
@@ -461,5 +453,20 @@ export function updateSchematicCollections(angularJson: Record<string, any>) {
   angularJson.cli.schematicCollections.unshift('@angular-eslint/schematics');
   // Delete old defaultCollection property if applicable
   delete angularJson.cli.defaultCollection;
+  return angularJson;
+}
+
+export function updateSchematicDefaults(
+  angularJson: Record<string, any>,
+  schematicFullName: string,
+  defaultValues: Record<string, unknown>,
+) {
+  angularJson.schematics = angularJson.schematics || {};
+  angularJson.schematics[schematicFullName] =
+    angularJson.schematics[schematicFullName] || {};
+  angularJson.schematics[schematicFullName] = {
+    ...angularJson.schematics[schematicFullName],
+    ...defaultValues,
+  };
   return angularJson;
 }